[scripts->utils] refactor
