+++ gaingrn/utils/bb_angle_tools.py
@@ -1,4 +1,4 @@
-## scripts/bb_angle_tools.py
+## utils/bb_angle_tools.py
 # Contains functions for handling backbone-angle related operations, which are used to find and denote backbone angle outliers in the STRIDE files.
 
 from cmath import phase, rect
