+++ gaingrn/utils/indexing_classes.py
@@ -1,10 +1,10 @@
-## scripts/indexing_classes.py
+## utils/indexing_classes.py
 # These are indexing classes for the GPCRDB-specific indexing of multiple GAIN domains.
 
 # Older classes like GPCRDBIndexing and Indexing are not used, since they were related to alignment-based indexing methods.
 
 import numpy as np
-import gaingrn.scripts.assign
+import gaingrn.utils.assign
 import glob
 import multiprocessing as mp
